mongoose => sequelize

+++ models/user.js
@@ -7,16 +7,14 @@ module.exports = (sequelize, DataTypes) => {
      * This method is not a part of Sequelize lifecycle.
      * The `models/index` file will call this method automatically.
      */
-    static associate(models) {
-      // define association here
-    }
+    static associate(models) {}
   }
   User.init(
     {
       userId: {
         primaryKey: true,
         autoIncrement: true,
-        type: DataTypes.BIGINT,
+        type: DataTypes.INTEGER,
       },
       email: DataTypes.STRING,
       nickname: DataTypes.STRING,
